LLM-code-generator/requirements.txt
fastapi==0.115.7
uvicorn[standard]==0.32.0

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# Pydantic V2
pydantic==2.10.4
pydantic-settings==2.6.1

# Auth & Utilities
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
email-validator==2.2.0

# Env & HTTP
python-dotenv==1.0.1
requests==2.32.3
python-multipart

# OpenAI API
openai==0.28

# Async performance (optional)
uvloop
